refactor: CategoryFilterSection 컴포넌트 생성 및 적용
- 테스트를 위해 카테고리 필터 선택 및 적용 부분을 감싸는 컴포넌트 생성
